Remove ActionPlugin as it's not used
Move ActionHandler inside the ActionModule and convert to a record.
